django_con/requirements.txt
django
psycopg2
djangorestframework
pyyaml
uritemplate